- client_svstag.h: change umodes field from unsigned int to uint64_t in ServicesTag struct

+++ src/client_svstag.h
@@ -31,7 +31,7 @@ struct ServicesTag
   list_node_t node;
   char *tag;
   unsigned int numeric;
-  unsigned int umodes;
+  uint64_t umodes;
 };
 
 extern void svstag_detach(list_t *, unsigned int);
